Swap height, width to make unletterbox measures consistent.
`x` is calculated to be "columns where the number of non-trivial
pixels is greater than 10% of the image height".  `y` is calculated
similarly, swapping "columns" for "rows" and "height" for "width".  If
both `x` and `y` are at least of length 2, then we use the first and
last entry of each as the bounds of the cropped image.  Otherwise, we
return the original image.

This PR corrected the height and width in the calculations for `x` and
`y`.